rename math function names

+++ examples/gray/pseudoShader.go
@@ -1,20 +1,5 @@
-// Write Your Shader in pseudoShader()
-//
-// resolution_x = 240
-// resolution_y = 160
-//
-// all variable type should be int16
-// ivec2 has x,y 
-// ivec4 has R,G,B,A
-//
-// R,G,B,A = 0..255, 0..255, 0..255, 0..255
-//
-// sin100 : sin(degree) with radius = 100
-// cos100 : cos(degree) with radius = 100
-// ilength(ivec2, ivec2) : euclidean distance ^2
-
 func pseudoShader() {
 	gray := int16(255/2)
 	half_transparent := int16(255/2)
-	gl_FragColor = ivec4{gray, gray, gray, half_transparent}
+	gl_FragColor = vec4{gray, gray, gray, half_transparent}
 }
